requirements.txt
streamlit
pandas
bcrypt
gspread
google-auth
google-api-python-client
oauth2client